requirements.txt
flask
requests
googlemaps
python-dotenv